Fix typo and add context of file to commit detail
well, i just answered one thanks to a stack overflow answer. I was
wondering how to use 'sub2ind' with a variable amount of subscripts...
the only thing i could think of was to generate a string of the actual
values at runtime, with commas between them, then use that in an eval
statement, but of course no one likes evals. turns out, there is a
simple way. use cell{:} and it will act like individual arguments. so
yay. the other question i have is this.  I'm trying to make a matrix
where the rows are the subscripts for each dimension for a point. the
next row is the next point. basically i'm trying to collect all the
subscripts that make up the middle 2 elements of an array with an
even-sized width or height. So an array of width 10 would middle element
positions of 5 and 6. A width 11 array would be 6 only.

So for a multidimentional array, say 2, or 3, or N, I want to collect
all the midpoint subscript positions (this is easy and I can do already)
but I want to lay them out so all the combinations are provided in a
matrix with each row as a point. If I have a 3-D array of dimension
10x10x10, the midpoints would be [5;6], [5;6], [5;6], but now i need to
lay them out so I have each combination.  2^3 = 8 total points so 8
rows.

My expected output is thus: [5, 5, 5; 5, 5, 6; 5, 6, 5; 5, 6, 6; 6, 5,
5; 6, 5, 6; 6, 6, 5; 6, 6, 6] The first column filled in a 4 and 4
pattern (8/2^1), second column in a 2 and 2 pattern (8/2^2), and third
in a 1 and 1 pattern (8/2^3).

I just cannot figure out a way (without using loops and counters.. which
I will do if i have to) to make this happen, using repmat or reshape or
anything. I need it programmatically so i can pass in any amount of
dimensions. Sometimes, one of the dimensions will have a single mid
point subscript because it was an odd-sized dimension... so for that
column, the single subscript would obviously fill it all the way down.
btw, happy father's day also, i got this new game called Striking
Vipers.. wanna play? this page
https://www.mathworks.com/matlabcentral/answers/333407-creating-matrix-of-given-pattern
has some interesting approaches but the particular example is pretty
different from mine. Theirs is much more repeatable. I think I solved
it:
>> reshape(reshape(kron([5;6]',[1;1;1;1]), 1, [])', [], 1) ans = 5 5 5 5
6 6 6 6

>> reshape(reshape(kron([5;6]',[1;1;1;1]), 2, [])', [], 1) ans = 5 5 6 6
5 5 6 6

>> reshape(reshape(kron([5;6]',[1;1;1;1]), 4, [])', [], 1) ans = 5 6 5 6
5 6 5 6

where the inputs 1,2,4 are related to the dimensions by 2^(dim-1) the
vector of ones can be made with ones(numOfPoints/2, 1) doesn't work if i
do a rand(10,9,10) oh well, i'll try again later. nah, i fixed it
already, just had to use a counter to replace the dim with the number of
times there has been a 2 element vector for a mid point. it works out
now.